Revert DB drop per test function with "Add tests for CoursePlan"
This reverts commit 7ab72c2aea92bdb9ca30c9423632016721fd6eba.

+++ tests/conftest.py
@@ -18,7 +18,6 @@ def setup(monkeypatch: pytest.MonkeyPatch):
     """
 
     mock_used = False
-
     def mock_get_db():
         nonlocal mock_used
         mock_used = True
@@ -31,17 +30,9 @@ def setup(monkeypatch: pytest.MonkeyPatch):
     monkeypatch.setattr("flaskr.db.course_plans.get_db", mock_get_db)
 
     yield
-    # Clean up the test database after tests (automatically)
-
-
-@pytest.fixture(autouse=True, scope="session")
-def database_setup():
-    """
-    Setup data of database.
-    This fixture is automatically applied to all tests and drop database after the whole test ends.
-    """
-    yield
-    get_mongo_client().drop_database(TEST_DB_NAME)
+    # Clean up the test database after tests
+    if mock_used:
+        get_mongo_client().drop_database(TEST_DB_NAME)
 
 
 @pytest.fixture
